sdm660: Remove useless memory dump region
This reserved memory dump region is intended to be used with the memory
dump v2 driver, but we've disabled that and we don't need this memory
dumping functionality. Remove the unused region and disable the
associated driver node to save memory.

* sdm660: Remove unused 36 MiB memdump region
This reserved memory dump region is intended to be used with the memory
dump v2 driver, but we've disabled that and we don't need this memory
dumping functionality. Remove the unused region and assiciated driver
node to save 36 MiB of memory.
Signed-off-by: Danny Lin <danny@kdrag0n.dev>

Change-Id: Ia0a307ee72822b33b21f4ba0a26ef224fcb7f2c5